32 bit int shifted by an amount not in the range -31..31 (FindBugs)
Java language specification:
"If the promoted type of the left-hand operand is int, only the five
lowest-order bits of the right-hand operand are used as the shift distance.
It is as if the right-hand operand were subjected to a bitwise logical AND
operator & (§15.22.1) with the mask value 0x1f.
The shift distance actually used is therefore always
in the range 0 to 31, inclusive."

+++ src/com/android/bluetooth/pbap/BluetoothPbapObexServer.java
@@ -507,14 +507,14 @@ public class BluetoothPbapObexServer extends ServerRequestHandler {
             switch (appParam[i]) {
                 case ApplicationParameter.TRIPLET_TAGID.FILTER_TAGID:
                     i += 2; // length and tag field in triplet
-                    appParamValue.filter = ((appParam[i+0] << 56) & 0xFF00000000000000L) |
-                                           ((appParam[i+1] << 48) & 0x00FF000000000000L) |
-                                           ((appParam[i+2] << 40) & 0x0000FF0000000000L) |
-                                           ((appParam[i+3] << 32) & 0x000000FF00000000L) |
-                                           ((appParam[i+4] << 24) & 0x00000000FF000000L) |
-                                           ((appParam[i+5] << 16) & 0x0000000000FF0000L) |
-                                           ((appParam[i+6] <<  8) & 0x000000000000FF00L) |
-                                           ((appParam[i+7] <<  0) & 0x00000000000000FFL);
+                    appParamValue.filter = (((long) appParam[i+0] << 56) & 0xFF00000000000000L) |
+                                           (((long) appParam[i+1] << 48) & 0x00FF000000000000L) |
+                                           (((long) appParam[i+2] << 40) & 0x0000FF0000000000L) |
+                                           (((long) appParam[i+3] << 32) & 0x000000FF00000000L) |
+                                           (((long) appParam[i+4] << 24) & 0x00000000FF000000L) |
+                                           (((long) appParam[i+5] << 16) & 0x0000000000FF0000L) |
+                                           (((long) appParam[i+6] <<  8) & 0x000000000000FF00L) |
+                                           (((long) appParam[i+7] <<  0) & 0x00000000000000FFL);
                     if (D) Log.i(TAG, "AppParam filter=" + appParamValue.filter);
                     i += ApplicationParameter.TRIPLET_LENGTH.FILTER_LENGTH;
                     break;
